AuthService/pyproject.toml
[project]
name = "authservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Nokotachi",email = "none@none.none"}
]
readme = "README.md"
requires-python = "=3.12.7"
dependencies = [
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "fastapi[all] (>=0.120.1,<0.121.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "alembic (>=1.17.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "authservice", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
